requirements.txt
torch==2.1.0
pyyaml
pydantic
attrs~=23.2.0
loguru~=0.7.2
tensorboard==2.4.0
matplotlib~=3.8.4
einops~=0.8.0
torchsummaryX~=1.3.0
jupyter
wandb~=0.17.0
numpy~=1.26.4
tqdm~=4.66.4
omegaconf~=2.3.0
ipython~=8.18.1
setuptools~=68.2.0
protobuf~=3.20